Improvements: User ID management & consider the CacheMemoryUser Adapter as an another domamin.
Manage the id within the user domain:
Compare to manage the id in the user adapters, it's better to manage it in the domain.
The advantage of this approach is that when we swtich database, we can keep the same id.

The cachedMemory could be both an adapter and a domain.